+++ packages/did-ui-react/src/components/ModalMethod/ManagerApprove/index.tsx
@@ -1,24 +1,20 @@
-import { Modal } from 'antd';
-import ManagerApproveContent, { ManagerApproveInnerProps } from './index.component';
-import './index.less';
-
+import { AuthServe } from '../../../utils/authServe';
+import ManagerApproveContent, { BaseManagerApproveInnerProps } from '../../ManagerApprove/index.component';
+import BaseModalFunc from '../BaseModalMethod';
 interface ManagerApproveModalProps {
   wrapClassName?: string;
   className?: string;
 }
 
-export type ManagerApproveProps = ManagerApproveInnerProps & ManagerApproveModalProps;
+export type ManagerApproveProps = BaseManagerApproveInnerProps & ManagerApproveModalProps;
 
 const managerApprove = async ({ wrapClassName, className, ...props }: ManagerApproveProps) =>
   new Promise((resolve, reject) => {
-    const modal = Modal.confirm({
-      width: 600,
-      icon: null,
-      centered: true,
-      okText: 'Confirm',
+    AuthServe.addRequestAuthCheck(props.originChainId);
+    const modal = BaseModalFunc({
       ...props,
-      wrapClassName: 'portkey-ui-wrapper portkey-ui-common-modals portkey-ui-manager-approve-wrapper ' + wrapClassName,
-      className: 'portkey-ui-manager-approve-modal ' + className,
+      wrapClassName: 'portkey-ui-manager-approve-modal-wrapper ' + wrapClassName,
+      className: 'portkey-ui-h-566 portkey-ui-manager-approve-modal ' + className,
       content: (
         <ManagerApproveContent
           {...props}
@@ -36,7 +32,6 @@ const managerApprove = async ({ wrapClassName, className, ...props }: ManagerApp
           }}
         />
       ),
-      okButtonProps: { style: { display: 'none' } },
     });
   });