+++ packages/services/src/service/services.ts
@@ -7,6 +7,7 @@ import {
   ITokenService,
   ITransactionService,
   IActivityService,
+  ISecurityService,
 } from '../types';
 import { IDIDGraphQL } from '@portkey/graphql';
 import { CommunityRecovery } from './communityRecovery';
@@ -15,6 +16,7 @@ import { Assets } from './assets';
 import { Token } from './token';
 import { Transaction } from './transaction';
 import { Activity } from './activity';
+import { Security } from './security';
 
 export class Services<T extends IBaseRequest = IBaseRequest> extends CommunityRecovery<T> implements IServices {
   readonly communityRecovery: ICommunityRecoveryService;
@@ -23,6 +25,7 @@ export class Services<T extends IBaseRequest = IBaseRequest> extends CommunityRe
   readonly token: ITokenService;
   readonly transaction: ITransactionService;
   readonly activity: IActivityService;
+  readonly security: ISecurityService;
 
   constructor(request: T, didGraphQL: IDIDGraphQL) {
     super(request, didGraphQL);
@@ -32,5 +35,6 @@ export class Services<T extends IBaseRequest = IBaseRequest> extends CommunityRe
     this.token = new Token(request);
     this.transaction = new Transaction(request);
     this.activity = new Activity(request);
+    this.security = new Security(request);
   }
 }